+++ test/modules/CloudToDeviceMessageTester/CloudToDeviceMessageTester.csproj
@@ -33,14 +33,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <None Remove="docker\linux\amd64\Dockerfile" />
-    <None Remove="docker\linux\arm32v7\Dockerfile" />
-    <None Remove="docker\linux\arm64v8\Dockerfile" />
-    <None Remove="docker\windows\amd64\Dockerfile" />
-    <None Remove="docker\windows\arm32v7\Dockerfile" />
-  </ItemGroup>
-
-  <ItemGroup>
     <PackageReference Include="Microsoft.Azure.Devices" Version="1.18.2" />
     <PackageReference Include="Microsoft.Azure.Devices.Client" Version="1.21.3" />
     <PackageReference Include="Microsoft.Extensions.Configuration" Version="2.2.0" />
